requirements.txt
# SKYWATCH Dependencies
# Core AI
insightface>=0.7.3              # Yüz algılama + tanıma (RetinaFace + ArcFace)
onnxruntime>=1.16.0             # Model çalıştırma (CPU)
onnxruntime-gpu                 # GPU varsa bunu etkinleştir

# Görüntü İşleme
opencv-python>=4.8.0            # Kamera yönetimi + frame işleme
numpy>=1.24.0                   # Sayısal hesaplama
scipy>=1.10.0                   # Cosine similarity
Pillow>=9.0                     # Görüntü yükleme/kaydetme

# Tracking
deep-sort-realtime>=1.3         # DeepSORT multi-object tracking

# GUI
PyQt5>=5.15.0                   # Masaüstü arayüz

# Config & Utils
pyyaml>=6.0                     # YAML config
